+++ src/Systems/AnimationSystem.hpp
@@ -15,9 +15,11 @@ public:
     ~AnimationSystem() override;
 
     const std::string getName() const override;
+
     void cleanUp() override;
 
-    void update(float deltaTime) override;
+
+    void update(int deltaTime) override;
 };
