+++ src/Systems/MovementSystem.hpp
@@ -14,9 +14,10 @@ public:
 
     ~MovementSystem() override;
 
-    void update(float deltaTime) override;
+    void update(int deltaTime) override;
 
     const std::string getName() const override;
+
     void cleanUp() override;
 };
