+++ src/Systems/TimerSystem.hpp
@@ -15,9 +15,10 @@ public:
     ~TimerSystem() override;
 
     const std::string getName() const override;
+
     void cleanUp() override;
 
-    void update(float deltaTime) override;
+    void update(int deltaTime) override;
 };
